requirements/requirements.txt
mysql.connector
fastapi
streamlit-extras
uvicorn
sqlalchemy
pandas 
numpy